+++ tests/spread/test_backup_pitr_aws.py/task.yaml
@@ -1,0 +1,9 @@
+summary: test_backup_pitr_aws.py
+environment:
+  TEST_MODULE: test_backup_pitr_aws.py
+execute: |
+  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
+backends:
+  - -lxd-vm  # Requires CI secrets
