+++ tests/spread/test_backup_pitr_gcp.py/task.yaml
@@ -1,6 +1,6 @@
-summary: test_backup_pitr.py
+summary: test_backup_pitr_gcp.py
 environment:
-  TEST_MODULE: test_backup_pitr.py
+  TEST_MODULE: test_backup_pitr_gcp.py
 execute: |
   tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
 artifacts:
